refactor(control-api): extract response_actions tests into sibling tests/
Move the inline #[cfg(test)] mod tests out of response_actions.rs into a
response_actions/tests/ subtree, split by theme into ack_receipts,
create_validation, delivery, and parse_ack. Shared fixtures
(test_response_action, test_ack, policy_rule_diff_action_payload,
policy_rule_diff_error_ack) live in tests/mod.rs as pub(super).

Pure relocation: no test logic changed; all 313 unit + 15 migration tests
pass. First step of the response_actions.rs god-file decomposition (step 7).
