+++ crates/services/control-api/src/routes/response_actions/mod.rs
@@ -2950,960 +2950,4 @@ async fn link_action_to_source_detection(
 }
 
 #[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod tests {
-    use super::*;
-
-    fn test_response_action(
-        target_kind: ResponseTargetKind,
-        action_type: &str,
-    ) -> ResponseActionRecord {
-        ResponseActionRecord {
-            id: Uuid::new_v4(),
-            tenant_id: Uuid::new_v4(),
-            action_type: action_type.to_string(),
-            target: ResponseTarget {
-                kind: target_kind,
-                id: "endpoint-1".to_string(),
-            },
-            requested_by: RequestedBy {
-                actor_type: "user".to_string(),
-                actor_id: "alice".to_string(),
-            },
-            requested_at: Utc::now(),
-            expires_at: None,
-            reason: "test".to_string(),
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: true,
-            payload: json!({}),
-            status: "published".to_string(),
-            metadata: json!({}),
-        }
-    }
-
-    fn test_ack(status: &str, target_kind: &str) -> AckSubmission {
-        parse_ack_submission(RecordResponseAckRequest {
-            target_kind: target_kind.to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: status.to_string(),
-            observed_at: None,
-            message: None,
-            resulting_state: None,
-            raw_payload: Some(json!({ "status": status })),
-        })
-        .expect("test acknowledgement parses")
-    }
-
-    fn policy_rule_diff_action_payload() -> Value {
-        json!({
-            "operation": "policy_rule_diff_validation",
-            "proposalId": "proposal-test",
-            "validationPlanSha256": "sha256:plan",
-            "endpointAgentId": "endpoint-1",
-            "request": {
-                "method": "POST",
-                "path": "/api/v1/edr/policy/event-impact",
-                "body": {
-                    "limit": 10,
-                    "trackPosture": true
-                }
-            }
-        })
-    }
-
-    fn policy_rule_diff_error_ack(mut error_payload: Value) -> AckSubmission {
-        error_payload["message"] = json!("local validation failed");
-        parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: "failed".to_string(),
-            observed_at: None,
-            message: Some("local validation failed".to_string()),
-            resulting_state: Some("policy_rule_diff_validation:failed".to_string()),
-            raw_payload: Some(json!({
-                "policyRuleDiffValidationError": error_payload
-            })),
-        })
-        .expect("test policy rule-diff failure acknowledgement parses")
-    }
-
-    #[test]
-    fn endpoint_terminal_ack_statuses_require_signed_receipts() {
-        let action = test_response_action(
-            ResponseTargetKind::Endpoint,
-            ResponseActionType::TerminateSession.as_str(),
-        );
-
-        for status in [
-            "acknowledged",
-            "rolled_back",
-            "failed",
-            "rejected",
-            "expired",
-        ] {
-            let ack = test_ack(status, "endpoint");
-            assert!(
-                requires_endpoint_ack_signed_receipt(&action, &ack),
-                "endpoint {status} acknowledgements must be signed"
-            );
-        }
-    }
-
-    #[test]
-    fn policy_rule_diff_non_ack_terminal_statuses_require_signed_receipts() {
-        let action = test_response_action(
-            ResponseTargetKind::Endpoint,
-            ResponseActionType::PolicyRuleDiffValidation.as_str(),
-        );
-
-        for status in ["failed", "rejected", "expired"] {
-            let ack = test_ack(status, "endpoint");
-            assert!(
-                requires_endpoint_ack_signed_receipt(&action, &ack),
-                "policy rule-diff {status} terminal acknowledgements must include an endpoint signed receipt"
-            );
-        }
-    }
-
-    #[test]
-    fn cloud_only_targets_do_not_require_endpoint_receipts() {
-        let action = test_response_action(
-            ResponseTargetKind::Principal,
-            ResponseActionType::RevokePrincipal.as_str(),
-        );
-        let ack = test_ack("acknowledged", "principal");
-
-        assert!(!requires_endpoint_ack_signed_receipt(&action, &ack));
-    }
-
-    #[test]
-    fn policy_rule_diff_error_payload_must_be_bound_to_dispatched_action() {
-        fn bad_request_message(err: ApiError) -> String {
-            match err {
-                ApiError::BadRequest(message) => message,
-                other => panic!("expected BadRequest, got {other:?}"),
-            }
-        }
-
-        let action_payload = policy_rule_diff_action_payload();
-        let empty_payload = parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: "failed".to_string(),
-            observed_at: None,
-            message: Some("local validation failed".to_string()),
-            resulting_state: Some("policy_rule_diff_validation:failed".to_string()),
-            raw_payload: Some(json!({})),
-        })
-        .expect("empty raw payload parses");
-        let message = bad_request_message(
-            validate_policy_rule_diff_error_payload_matches_action(
-                &empty_payload,
-                &action_payload,
-                "endpoint-1",
-            )
-            .unwrap_err(),
-        );
-        assert!(message.contains("must include policyRuleDiffValidationError"));
-
-        let wrong_plan = policy_rule_diff_error_ack(json!({
-            "proposalId": "proposal-test",
-            "validationPlanSha256": "sha256:wrong-plan",
-            "endpointAgentId": "endpoint-1",
-            "request": action_payload["request"].clone()
-        }));
-        let message = bad_request_message(
-            validate_policy_rule_diff_error_payload_matches_action(
-                &wrong_plan,
-                &action_payload,
-                "endpoint-1",
-            )
-            .unwrap_err(),
-        );
-        assert!(message.contains("validationPlanSha256 does not match"));
-
-        let wrong_request = policy_rule_diff_error_ack(json!({
-            "proposalId": "proposal-test",
-            "validationPlanSha256": "sha256:plan",
-            "endpointAgentId": "endpoint-1",
-            "request": {
-                "method": "POST",
-                "path": "/api/v1/edr/policy/event-impact",
-                "body": {
-                    "limit": 999,
-                    "trackPosture": true
-                }
-            }
-        }));
-        let message = bad_request_message(
-            validate_policy_rule_diff_error_payload_matches_action(
-                &wrong_request,
-                &action_payload,
-                "endpoint-1",
-            )
-            .unwrap_err(),
-        );
-        assert!(message.contains("request does not match"));
-
-        let valid = policy_rule_diff_error_ack(json!({
-            "proposalId": "proposal-test",
-            "validationPlanSha256": "sha256:plan",
-            "endpointAgentId": "endpoint-1",
-            "request": action_payload["request"].clone()
-        }));
-        validate_policy_rule_diff_error_payload_matches_action(
-            &valid,
-            &action_payload,
-            "endpoint-1",
-        )
-        .expect("bound policy rule-diff error payload should validate");
-    }
-
-    #[test]
-    fn delivery_plan_uses_response_subject_for_supported_endpoint_actions() {
-        let action = ResponseActionRecord {
-            id: Uuid::new_v4(),
-            tenant_id: Uuid::new_v4(),
-            action_type: "policy_rule_diff_validation".to_string(),
-            target: ResponseTarget {
-                kind: ResponseTargetKind::Endpoint,
-                id: "agent-123".to_string(),
-            },
-            requested_by: RequestedBy {
-                actor_type: "user".to_string(),
-                actor_id: "alice".to_string(),
-            },
-            requested_at: Utc::now(),
-            expires_at: None,
-            reason: "test".to_string(),
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: true,
-            payload: json!({"operation": "policy_rule_diff_validation"}),
-            status: "queued".to_string(),
-            metadata: json!({}),
-        };
-
-        let plan = delivery_plan(&action, "acme");
-        assert_eq!(
-            plan.delivery_subject.as_deref(),
-            Some("tenant-acme.clawdstrike.response.command.endpoint.agent-123")
-        );
-        assert!(plan.metadata["compat_mirror_subject"].is_null());
-        assert_eq!(
-            plan.metadata["canonical_subject"],
-            "tenant-acme.clawdstrike.response.command.endpoint.agent-123"
-        );
-        assert!(plan.metadata["ack_token"].is_string());
-    }
-
-    #[test]
-    fn publish_contract_rejects_legacy_endpoint_actions() {
-        let action = ResponseActionRecord {
-            id: Uuid::new_v4(),
-            tenant_id: Uuid::new_v4(),
-            action_type: "request_policy_reload".to_string(),
-            target: ResponseTarget {
-                kind: ResponseTargetKind::Endpoint,
-                id: "agent-123".to_string(),
-            },
-            requested_by: RequestedBy {
-                actor_type: "user".to_string(),
-                actor_id: "alice".to_string(),
-            },
-            requested_at: Utc::now(),
-            expires_at: None,
-            reason: "reload".to_string(),
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: false,
-            payload: json!({}),
-            status: "queued".to_string(),
-            metadata: json!({}),
-        };
-
-        let err = validate_publish_delivery_contract(&action).unwrap_err();
-        match err {
-            ApiError::BadRequest(message) => {
-                assert!(message.contains("policy_rule_diff_validation"));
-            }
-            other => panic!("expected BadRequest, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn cloud_only_targets_skip_transport_subject() {
-        let action = ResponseActionRecord {
-            id: Uuid::new_v4(),
-            tenant_id: Uuid::new_v4(),
-            action_type: "revoke_principal".to_string(),
-            target: ResponseTarget {
-                kind: ResponseTargetKind::Principal,
-                id: "principal-1".to_string(),
-            },
-            requested_by: RequestedBy {
-                actor_type: "user".to_string(),
-                actor_id: "alice".to_string(),
-            },
-            requested_at: Utc::now(),
-            expires_at: None,
-            reason: "test".to_string(),
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: false,
-            payload: json!({}),
-            status: "queued".to_string(),
-            metadata: json!({}),
-        };
-
-        let plan = delivery_plan(&action, "acme");
-        assert!(plan.delivery_subject.is_none());
-        assert_eq!(plan.executor_kind, "cloud_only");
-    }
-
-    #[test]
-    fn create_validation_rejects_invalid_action_target_pairs() {
-        let input = CreateResponseActionRequest {
-            action_type: "request_policy_reload".to_string(),
-            target: ResponseTargetInput {
-                kind: "principal".to_string(),
-                id: "p-1".to_string(),
-            },
-            reason: "reload".to_string(),
-            expires_at: None,
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: Some(false),
-            payload: None,
-        };
-        let err = validate_create_request(
-            &input,
-            &ResponseActionType::RequestPolicyReload,
-            &ResponseTargetKind::Principal,
-            false,
-        )
-        .unwrap_err();
-        assert!(matches!(err, ApiError::BadRequest(_)));
-    }
-
-    #[test]
-    fn create_validation_rejects_endpoint_actions_without_agent_executor() {
-        let input = CreateResponseActionRequest {
-            action_type: "request_policy_reload".to_string(),
-            target: ResponseTargetInput {
-                kind: "endpoint".to_string(),
-                id: "agent-1".to_string(),
-            },
-            reason: "reload".to_string(),
-            expires_at: None,
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: Some(false),
-            payload: None,
-        };
-
-        let err = validate_create_request(
-            &input,
-            &ResponseActionType::RequestPolicyReload,
-            &ResponseTargetKind::Endpoint,
-            false,
-        )
-        .unwrap_err();
-        match err {
-            ApiError::BadRequest(message) => {
-                assert!(message.contains("policy_rule_diff_validation"));
-            }
-            other => panic!("expected BadRequest, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn create_validation_requires_policy_rule_diff_acknowledgement() {
-        let input = CreateResponseActionRequest {
-            action_type: "policy_rule_diff_validation".to_string(),
-            target: ResponseTargetInput {
-                kind: "endpoint".to_string(),
-                id: "agent-1".to_string(),
-            },
-            reason: "validate proposal".to_string(),
-            expires_at: None,
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: Some(false),
-            payload: Some(json!({"operation": "policy_rule_diff_validation"})),
-        };
-
-        let err = validate_create_request(
-            &input,
-            &ResponseActionType::PolicyRuleDiffValidation,
-            &ResponseTargetKind::Endpoint,
-            false,
-        )
-        .unwrap_err();
-        match err {
-            ApiError::BadRequest(message) => {
-                assert!(message.contains("require acknowledgement"));
-            }
-            other => panic!("expected BadRequest, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn create_validation_accepts_policy_rule_diff_endpoint_contract() {
-        let input = CreateResponseActionRequest {
-            action_type: "policy_rule_diff_validation".to_string(),
-            target: ResponseTargetInput {
-                kind: "endpoint".to_string(),
-                id: "agent-1".to_string(),
-            },
-            reason: "validate proposal".to_string(),
-            expires_at: Some(Utc::now() + Duration::minutes(5)),
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: Some(true),
-            payload: Some(json!({"operation": "policy_rule_diff_validation"})),
-        };
-
-        validate_create_request(
-            &input,
-            &ResponseActionType::PolicyRuleDiffValidation,
-            &ResponseTargetKind::Endpoint,
-            true,
-        )
-        .expect("policy_rule_diff_validation endpoint command should be accepted");
-    }
-
-    #[test]
-    fn transition_posture_requires_target_state_in_payload() {
-        let input = CreateResponseActionRequest {
-            action_type: "transition_posture".to_string(),
-            target: ResponseTargetInput {
-                kind: "project".to_string(),
-                id: "agent-1".to_string(),
-            },
-            reason: "contain".to_string(),
-            expires_at: None,
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: Some(false),
-            payload: Some(json!({})),
-        };
-
-        let err = validate_create_request(
-            &input,
-            &ResponseActionType::TransitionPosture,
-            &ResponseTargetKind::Project,
-            false,
-        )
-        .unwrap_err();
-        match err {
-            ApiError::BadRequest(message) => {
-                assert!(message.contains("payload.toState"));
-            }
-            other => panic!("expected BadRequest, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn create_validation_rejects_acknowledgement_until_executor_support_exists() {
-        let input = CreateResponseActionRequest {
-            action_type: "request_policy_reload".to_string(),
-            target: ResponseTargetInput {
-                kind: "endpoint".to_string(),
-                id: "agent-1".to_string(),
-            },
-            reason: "reload".to_string(),
-            expires_at: None,
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: Some(true),
-            payload: None,
-        };
-
-        let err = validate_create_request(
-            &input,
-            &ResponseActionType::RequestPolicyReload,
-            &ResponseTargetKind::Endpoint,
-            true,
-        )
-        .unwrap_err();
-        assert!(matches!(err, ApiError::BadRequest(_)));
-    }
-
-    #[test]
-    fn create_validation_rejects_unsupported_runtime_targets() {
-        let input = CreateResponseActionRequest {
-            action_type: "kill_switch".to_string(),
-            target: ResponseTargetInput {
-                kind: "runtime".to_string(),
-                id: "runtime-1".to_string(),
-            },
-            reason: "contain".to_string(),
-            expires_at: None,
-            case_id: None,
-            source_detection_id: None,
-            source_approval_id: None,
-            require_acknowledgement: Some(false),
-            payload: None,
-        };
-
-        let err = validate_create_request(
-            &input,
-            &ResponseActionType::KillSwitch,
-            &ResponseTargetKind::Runtime,
-            false,
-        )
-        .unwrap_err();
-        assert!(matches!(err, ApiError::BadRequest(_)));
-    }
-
-    #[test]
-    fn create_validation_rejects_oversized_request_fields() {
-        let oversized_target = match validate_create_request(
-            &CreateResponseActionRequest {
-                action_type: "request_policy_reload".to_string(),
-                target: ResponseTargetInput {
-                    kind: "endpoint".to_string(),
-                    id: "e".repeat(257),
-                },
-                reason: "reload".to_string(),
-                expires_at: None,
-                case_id: None,
-                source_detection_id: None,
-                source_approval_id: None,
-                require_acknowledgement: Some(false),
-                payload: None,
-            },
-            &ResponseActionType::RequestPolicyReload,
-            &ResponseTargetKind::Endpoint,
-            false,
-        ) {
-            Ok(_) => panic!("oversized response-action target id was accepted"),
-            Err(err) => err,
-        };
-        assert!(matches!(oversized_target, ApiError::BadRequest(_)));
-
-        let oversized_reason = match validate_create_request(
-            &CreateResponseActionRequest {
-                action_type: "request_policy_reload".to_string(),
-                target: ResponseTargetInput {
-                    kind: "endpoint".to_string(),
-                    id: "agent-1".to_string(),
-                },
-                reason: "r".repeat(2049),
-                expires_at: None,
-                case_id: None,
-                source_detection_id: None,
-                source_approval_id: None,
-                require_acknowledgement: Some(false),
-                payload: None,
-            },
-            &ResponseActionType::RequestPolicyReload,
-            &ResponseTargetKind::Endpoint,
-            false,
-        ) {
-            Ok(_) => panic!("oversized response-action reason was accepted"),
-            Err(err) => err,
-        };
-        assert!(matches!(oversized_reason, ApiError::BadRequest(_)));
-
-        let oversized_payload = match validate_create_request(
-            &CreateResponseActionRequest {
-                action_type: "request_policy_reload".to_string(),
-                target: ResponseTargetInput {
-                    kind: "endpoint".to_string(),
-                    id: "agent-1".to_string(),
-                },
-                reason: "reload".to_string(),
-                expires_at: None,
-                case_id: None,
-                source_detection_id: None,
-                source_approval_id: None,
-                require_acknowledgement: Some(false),
-                payload: Some(json!({
-                    "blob": "p".repeat(65_536)
-                })),
-            },
-            &ResponseActionType::RequestPolicyReload,
-            &ResponseTargetKind::Endpoint,
-            false,
-        ) {
-            Ok(_) => panic!("oversized response-action payload was accepted"),
-            Err(err) => err,
-        };
-        assert!(matches!(oversized_payload, ApiError::BadRequest(_)));
-    }
-
-    #[test]
-    fn control_discriminator_errors_do_not_echo_oversized_values() {
-        fn bad_request_message(err: ApiError) -> String {
-            match err {
-                ApiError::BadRequest(message) => message,
-                other => panic!("expected BadRequest, got {other:?}"),
-            }
-        }
-
-        let oversized_action_type = "a".repeat(2048);
-        let action_message =
-            bad_request_message(ResponseActionType::from_str(&oversized_action_type).unwrap_err());
-        assert!(!action_message.contains(&oversized_action_type));
-        assert!(action_message.len() < 256);
-
-        let oversized_target_kind = "t".repeat(2048);
-        let target_message =
-            bad_request_message(ResponseTargetKind::from_str(&oversized_target_kind).unwrap_err());
-        assert!(!target_message.contains(&oversized_target_kind));
-        assert!(target_message.len() < 256);
-
-        let oversized_ack_status = "s".repeat(2048);
-        let status_message =
-            bad_request_message(normalize_ack_status(&oversized_ack_status).unwrap_err());
-        assert!(!status_message.contains(&oversized_ack_status));
-        assert!(status_message.len() < 256);
-    }
-
-    #[test]
-    fn control_discriminators_reject_oversized_values_with_length_errors() {
-        fn bad_request_message(err: ApiError) -> String {
-            match err {
-                ApiError::BadRequest(message) => message,
-                other => panic!("expected BadRequest, got {other:?}"),
-            }
-        }
-
-        let action_message =
-            bad_request_message(ResponseActionType::from_str(&"a".repeat(65)).unwrap_err());
-        assert!(action_message.contains("action_type"));
-        assert!(action_message.contains("at most"));
-
-        let target_message =
-            bad_request_message(ResponseTargetKind::from_str(&"t".repeat(65)).unwrap_err());
-        assert!(target_message.contains("target.kind"));
-        assert!(target_message.contains("at most"));
-
-        let status_message =
-            bad_request_message(normalize_ack_status(&"s".repeat(65)).unwrap_err());
-        assert!(status_message.contains("status"));
-        assert!(status_message.contains("at most"));
-    }
-
-    #[test]
-    fn response_action_requests_reject_unknown_fields() {
-        let create_err = serde_json::from_value::<CreateResponseActionRequest>(json!({
-            "actionType": "request_policy_reload",
-            "target": {
-                "kind": "endpoint",
-                "id": "agent-1"
-            },
-            "reason": "reload",
-            "dryRun": true
-        }))
-        .expect_err("unknown create request field should be rejected");
-        assert!(create_err.to_string().contains("unknown field"));
-
-        let target_err = serde_json::from_value::<CreateResponseActionRequest>(json!({
-            "actionType": "request_policy_reload",
-            "target": {
-                "kind": "endpoint",
-                "id": "agent-1",
-                "displayName": "Endpoint One"
-            },
-            "reason": "reload"
-        }))
-        .expect_err("unknown target field should be rejected");
-        assert!(target_err.to_string().contains("unknown field"));
-
-        let ack_err = serde_json::from_value::<RecordResponseAckRequest>(json!({
-            "targetKind": "endpoint",
-            "targetId": "agent-1",
-            "ackToken": "ack-token",
-            "status": "acknowledged",
-            "receiptHash": "abc123"
-        }))
-        .expect_err("unknown acknowledgement field should be rejected");
-        assert!(ack_err.to_string().contains("unknown field"));
-    }
-
-    #[test]
-    fn parse_ack_submission_rejects_untrusted_observed_at() {
-        let future_observed_at = match parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: "acknowledged".to_string(),
-            observed_at: Some(Utc::now() + Duration::minutes(6)),
-            message: None,
-            resulting_state: None,
-            raw_payload: None,
-        }) {
-            Ok(_) => panic!("future acknowledgement observed_at was accepted"),
-            Err(err) => err,
-        };
-        assert!(matches!(future_observed_at, ApiError::BadRequest(_)));
-
-        let stale_observed_at = match parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: "acknowledged".to_string(),
-            observed_at: Some(Utc::now() - Duration::hours(2)),
-            message: None,
-            resulting_state: None,
-            raw_payload: None,
-        }) {
-            Ok(_) => panic!("stale acknowledgement observed_at was accepted"),
-            Err(err) => err,
-        };
-        assert!(matches!(stale_observed_at, ApiError::BadRequest(_)));
-    }
-
-    #[test]
-    fn normalize_ack_status_accepts_rollback_acknowledgement() {
-        assert_eq!(normalize_ack_status("rolled_back").unwrap(), "rolled_back");
-    }
-
-    #[test]
-    fn parse_ack_submission_rejects_oversized_ack_fields() {
-        let oversized_target = match parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "e".repeat(257),
-            ack_token: "ack-token".to_string(),
-            status: "acknowledged".to_string(),
-            observed_at: None,
-            message: None,
-            resulting_state: None,
-            raw_payload: None,
-        }) {
-            Ok(_) => panic!("oversized acknowledgement target id was accepted"),
-            Err(err) => err,
-        };
-        assert!(matches!(oversized_target, ApiError::BadRequest(_)));
-
-        let oversized_token = match parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "t".repeat(1025),
-            status: "acknowledged".to_string(),
-            observed_at: None,
-            message: None,
-            resulting_state: None,
-            raw_payload: None,
-        }) {
-            Ok(_) => panic!("oversized acknowledgement token was accepted"),
-            Err(err) => err,
-        };
-        assert!(matches!(oversized_token, ApiError::BadRequest(_)));
-
-        let oversized_message = match parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: "acknowledged".to_string(),
-            observed_at: None,
-            message: Some("m".repeat(2049)),
-            resulting_state: None,
-            raw_payload: None,
-        }) {
-            Ok(_) => panic!("oversized acknowledgement message was accepted"),
-            Err(err) => err,
-        };
-        assert!(matches!(oversized_message, ApiError::BadRequest(_)));
-
-        let oversized_resulting_state = match parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: "acknowledged".to_string(),
-            observed_at: None,
-            message: None,
-            resulting_state: Some("s".repeat(257)),
-            raw_payload: None,
-        }) {
-            Ok(_) => panic!("oversized acknowledgement resulting state was accepted"),
-            Err(err) => err,
-        };
-        assert!(matches!(oversized_resulting_state, ApiError::BadRequest(_)));
-
-        let oversized_raw_payload = match parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: "acknowledged".to_string(),
-            observed_at: None,
-            message: None,
-            resulting_state: None,
-            raw_payload: Some(json!({
-                "blob": "r".repeat(65_536)
-            })),
-        }) {
-            Ok(_) => panic!("oversized acknowledgement raw payload was accepted"),
-            Err(err) => err,
-        };
-        assert!(matches!(oversized_raw_payload, ApiError::BadRequest(_)));
-    }
-
-    #[test]
-    fn policy_rule_diff_ack_payload_requires_receipt_and_impact_for_success() {
-        fn bad_request_message(err: ApiError) -> String {
-            match err {
-                ApiError::BadRequest(message) => message,
-                other => panic!("expected BadRequest, got {other:?}"),
-            }
-        }
-
-        let missing_receipt = parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: "acknowledged".to_string(),
-            observed_at: None,
-            message: None,
-            resulting_state: None,
-            raw_payload: Some(json!({
-                "policyRuleDiffValidation": {
-                    "impact": {
-                        "impactId": "impact-test"
-                    }
-                }
-            })),
-        })
-        .expect("ack payload parses");
-        let message = bad_request_message(
-            validate_policy_rule_diff_ack_payload(&missing_receipt).unwrap_err(),
-        );
-        assert!(message.contains("policyRuleDiffValidation acknowledgement must include receipt"));
-
-        let complete = parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: "acknowledged".to_string(),
-            observed_at: None,
-            message: None,
-            resulting_state: None,
-            raw_payload: Some(json!({
-                "policyRuleDiffValidation": {
-                    "impact": {
-                        "impactId": "impact-test"
-                    },
-                    "receipt": {
-                        "receipt": {
-                            "receipt_id": "receipt-test"
-                        }
-                    }
-                }
-            })),
-        })
-        .expect("ack payload parses");
-        validate_policy_rule_diff_ack_payload(&complete)
-            .expect("complete policy rule-diff acknowledgement payload");
-
-        let failed_without_receipt = parse_ack_submission(RecordResponseAckRequest {
-            target_kind: "endpoint".to_string(),
-            target_id: "endpoint-1".to_string(),
-            ack_token: "ack-token".to_string(),
-            status: "failed".to_string(),
-            observed_at: None,
-            message: Some("local validation failed".to_string()),
-            resulting_state: None,
-            raw_payload: Some(json!({
-                "policyRuleDiffValidationError": {
-                    "message": "local validation failed"
-                }
-            })),
-        })
-        .expect("ack payload parses");
-        validate_policy_rule_diff_ack_payload(&failed_without_receipt).expect(
-            "error payload shape validation is separate from terminal signed-receipt enforcement",
-        );
-    }
-
-    #[test]
-    fn policy_rule_diff_ack_binds_dispatched_expected_receipt() {
-        fn bad_request_message(err: ApiError) -> String {
-            match err {
-                ApiError::BadRequest(message) => message,
-                other => panic!("expected BadRequest, got {other:?}"),
-            }
-        }
-
-        let expected_receipt = json!({
-            "receiptFamily": "policy_event_impact",
-            "ruleId": "endpoint.policy_event.impact",
-            "graphProcessNodeId": "fleet-rule-diff",
-            "proposedPolicyHash": "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-            "proposedPolicyEpoch": 42,
-            "requiredEvidenceKeys": [
-                "impactId",
-                "proposedPolicyHash",
-                "proposedPolicyEpoch"
-            ]
-        });
-        let action_payload = json!({
-            "expectedReceipt": expected_receipt.clone()
-        });
-        let payload = json!({
-            "endpointAgentId": "agent-1",
-            "expectedReceipt": expected_receipt,
-            "impact": {
-                "proposedPolicyHash": "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "proposedPolicyEpoch": 42
-            }
-        });
-        validate_policy_rule_diff_ack_matches_action_payload(&payload, &action_payload, "agent-1")
-            .expect("expected receipt identity should match");
-
-        let mut mutated_expected = payload.clone();
-        mutated_expected["expectedReceipt"]["proposedPolicyHash"] =
-            json!("0xbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb");
-        let message = bad_request_message(
-            validate_policy_rule_diff_ack_matches_action_payload(
-                &mutated_expected,
-                &action_payload,
-                "agent-1",
-            )
-            .unwrap_err(),
-        );
-        assert!(message.contains("expectedReceipt does not match"));
-
-        let mut mutated_impact = payload.clone();
-        mutated_impact["impact"]["proposedPolicyEpoch"] = json!(43);
-        let message = bad_request_message(
-            validate_policy_rule_diff_ack_matches_action_payload(
-                &mutated_impact,
-                &action_payload,
-                "agent-1",
-            )
-            .unwrap_err(),
-        );
-        assert!(message.contains("impact proposedPolicyEpoch"));
-    }
-
-    #[test]
-    fn scrub_delivery_metadata_hides_ack_token() {
-        let scrubbed = scrub_delivery_metadata(json!({
-            "ack_token": "secret",
-            "compat_mirror_subject": "tenant-acme.clawdstrike.posture.command.agent-123",
-        }));
-
-        assert!(scrubbed.get("ack_token").is_none());
-        assert_eq!(
-            scrubbed["compat_mirror_subject"],
-            "tenant-acme.clawdstrike.posture.command.agent-123"
-        );
-    }
-}
+mod tests;
